ci: Go 1.11.1 support
Cherry pick origin 0fee806ad4970060e8f97d37a50fc58572426195 for Go
1.11.1 support.
