+++ compiler/prelude/genprelude.go
@@ -5,6 +5,7 @@ package main
 import (
 	"bytes"
 	"fmt"
+	"go/build"
 	"io/ioutil"
 	"log"
 	"os/exec"
@@ -24,17 +25,12 @@ func main() {
 }
 
 func run() error {
-	var bpkgDir string
-	{
-		cmd := exec.Command("go", "list", "-m", "-f", "{{.Dir}}")
-		out, err := cmd.Output()
-		if err != nil {
-			return fmt.Errorf("failed to resolve module path: %v", err)
-		}
-		bpkgDir = strings.TrimSpace(string(out))
+	bpkg, err := build.Import("github.com/gopherjs/gopherjs", ".", build.FindOnly)
+	if err != nil {
+		return fmt.Errorf("failed to locate path for github.com/gopherjs/gopherjs: %v", err)
 	}
 
-	preludeDir := filepath.Join(bpkgDir, "compiler", "prelude")
+	preludeDir := filepath.Join(bpkg.Dir, "compiler", "prelude")
 
 	files := []string{
 		"prelude.js",
@@ -58,7 +54,7 @@ func run() error {
 	}
 
 	args := append([]string{
-		filepath.Join(bpkgDir, "node_modules", ".bin", "uglifyjs"),
+		filepath.Join(bpkg.Dir, "node_modules", ".bin", "uglifyjs"),
 		"--config-file",
 		filepath.Join(preludeDir, "uglifyjs_options.json"),
 	}, files...)
