+++ compiler/version_check.go
@@ -6,4 +6,4 @@ package compiler
 const ___GOPHERJS_REQUIRES_GO_VERSION_1_11___ = true
 
 // Version is the GopherJS compiler version string.
-const Version = "1.11"
+const Version = "1.11-2"
