+++ go.mod
@@ -7,7 +7,7 @@ require (
 	github.com/neelance/sourcemap v0.0.0-20151028013722-8c68805598ab
 	github.com/shurcooL/go v0.0.0-20180423040247-9e1955d9fb6e
 	github.com/shurcooL/httpfs v0.0.0-20171119174359-809beceb2371
-	github.com/shurcooL/vfsgen v0.0.0-20180711163814-62bca832be04
+	github.com/shurcooL/vfsgen v0.0.0-20180915214035-33ae1944be3f
 	github.com/spf13/cobra v0.0.3
 	github.com/spf13/pflag v1.0.1
 	golang.org/x/crypto v0.0.0-20180807104621-f027049dab0a
